+++ app/modules/custom_reports/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.3.0'
+__version__ = '0.4.0'
 
 from .routes import custom_reports_bp
 from .service import CustomReportsService