+++ app/modules/custom_reports/queries.py
@@ -24,7 +24,8 @@ class CustomReportsQueries:
             v.vend_name AS brand,
             i.item_size AS item_size,
             i.UDF5_STRING AS season,
-            NVL(iq.qty, 0) AS on_hand_qty
+            NVL(iq.qty, 0) AS on_hand_qty,
+            i.first_rcvd_date AS first_rcvd_date
         FROM rps.invn_sbs_item i
         LEFT JOIN rps.vendor v ON v.sid = i.vend_sid
         LEFT JOIN (